chore: tighten install path integrity (.old swap, dpkg-query fail-closed, file: secret cleanup) (#1360) (#1384)
I tightened install path integrity across the shell and Rust update
paths. The change preserves target modes from private staging, unifies
Unix and Windows recovery on `target.old`, cleans podup-created native
secrets during `down`, and makes the `dpkg-query` ownership path skip
manifest scanning when the helper is unavailable.

Closes #1360

Validation: `cargo build --locked`, `cargo test --locked`, `shellcheck
install.sh`. PSScriptAnalyzer was not run because `pwsh` is unavailable
in this environment. The Podman 5 and 6 lane remains covered by CI.
