+++ internal/engine/secrets/remove.rs
@@ -4,6 +4,16 @@
 //! secret (#1263). That list is also a superset of what the compose file names,
 //! since every secret podup creates carries `podup.project=<proj>`, so a single
 //! pass over it sweeps the declared secrets and the orphans alike.
+//!
+//! Every secret podup creates — including `file:` sources, whose bytes are
+//! copied into a Podman-native secret at `up` time and **persist independently
+//! of any container** (`podman-secret-rm(1)`: the secret lives in the podman
+//! store, not in container state) — is removed on `down`. Without this sweep
+//! the bytes from a `file:` source survive a `down`/`up` cycle until the next
+//! `up` overwrites them, which is a defence-in-depth gap (#1360): a stale
+//! secret payload outlives the container that read it. A summary log line
+//! names the secrets that were removed, so the operator can see the teardown
+//! actually happened.
 
 use crate::compose::types::ComposeFile;
 use crate::error::Result;
@@ -35,6 +45,11 @@ impl Engine {
 		// orphans (a key since renamed, or a `down` run without the original
 		// file) in one pass. A same-named secret the user created by hand is not
 		// in the set, which is exactly the guard this has to keep.
+		//
+		// Both paths emit a single summary log line so the operator can see
+		// which secrets were removed (#1360 L6). The per-secret `info` lines
+		// stay; the summary is the closure the user is actually looking for
+		// when the teardown finishes.
 		match self.list_project_secret_names().await {
 			Some(owned) => {
 				// Ownership is already established by the label on each entry, so
@@ -43,9 +58,13 @@ impl Engine {
 				// is best-effort here either way (a delete failure is logged, not
 				// fatal), and a secret that changed hands inside it would have been
 				// created by podup moments earlier.
+				let mut removed = Vec::with_capacity(owned.len());
 				for name in owned {
-					self.delete_listed_secret(&name).await;
+					if self.delete_listed_secret(&name).await {
+						removed.push(name);
+					}
 				}
+				log_removed_secrets(&self.project, &removed);
 			}
 			// The list failed. Falling through to an empty set would silently
 			// delete nothing and report a clean teardown, so this drops back to
@@ -54,6 +73,7 @@ impl Engine {
 			// sweep is not possible without a list, which is also how it behaved
 			// before.
 			None => {
+				let mut removed = Vec::new();
 				for (name, def) in &file.secrets {
 					if is_podup_created_source(
 						def.external,
@@ -61,8 +81,10 @@ impl Engine {
 						def.environment.as_deref(),
 						def.file.as_deref(),
 					) {
-						self.delete_secret(&scoped_name(&self.project, "secret", name))
-							.await;
+						let scoped = scoped_name(&self.project, "secret", name);
+						if self.delete_secret(&scoped).await {
+							removed.push(scoped);
+						}
 					}
 				}
 				for (name, def) in &file.configs {
@@ -72,10 +94,13 @@ impl Engine {
 						def.environment.as_deref(),
 						def.file.as_deref(),
 					) {
-						self.delete_secret(&scoped_name(&self.project, "config", name))
-							.await;
+						let scoped = scoped_name(&self.project, "config", name);
+						if self.delete_secret(&scoped).await {
+							removed.push(scoped);
+						}
 					}
 				}
+				log_removed_secrets(&self.project, &removed);
 			}
 		}
 		Ok(())
@@ -127,11 +152,21 @@ impl Engine {
 	/// own. Kept separate from [`Self::delete_secret`] rather than adding a flag,
 	/// so that a call site can never accidentally skip a check it was supposed to
 	/// make: this one is only reachable from a name the list vouched for.
-	async fn delete_listed_secret(&self, name: &str) {
+	///
+	/// Returns `true` when the secret was actually removed (or was already
+	/// absent — a 404 still counts as teardown success). The caller collects
+	/// the names for the summary log line.
+	async fn delete_listed_secret(&self, name: &str) -> bool {
 		let path = format!("{API_PREFIX}/secrets/{}", urlencoded(name));
 		match self.client.delete_ok(&path).await {
-			Ok(()) => tracing::info!("removed secret {name}"),
-			Err(e) => tracing::warn!("could not remove secret {name}: {e}"),
+			Ok(()) => {
+				tracing::info!("removed secret {name}");
+				true
+			}
+			Err(e) => {
+				tracing::warn!("could not remove secret {name}: {e}");
+				false
+			}
 		}
 	}
 
@@ -139,7 +174,11 @@ impl Engine {
 	/// `podup.project=<proj>` label — so a same-named secret the user created by
 	/// hand (and which podup never created) is never destroyed on `down`. A
 	/// missing secret (404) is a no-op.
-	async fn delete_secret(&self, name: &str) {
+	///
+	/// Returns `true` when the secret was actually removed, `false` when the
+	/// inspect failed (404, transport error, or a foreign label) and the
+	/// caller should not count it as teardown.
+	async fn delete_secret(&self, name: &str) -> bool {
 		let inspect = format!("{API_PREFIX}/secrets/{}/json", urlencoded(name));
 		match self.client.get_json::<serde_json::Value>(&inspect).await {
 			Ok(info) => {
@@ -155,21 +194,44 @@ impl Engine {
 						 leaving it untouched (not created by podup)",
 						self.project
 					);
-					return;
+					return false;
 				}
 			}
-			Err(e) if e.is_status(404) => return,
+			Err(e) if e.is_status(404) => return false,
 			Err(e) => {
 				tracing::warn!("could not inspect secret {name} before removal: {e}");
-				return;
+				return false;
 			}
 		}
 		let path = format!("{API_PREFIX}/secrets/{}", urlencoded(name));
 		match self.client.delete_ok(&path).await {
-			Ok(()) => tracing::info!("removed secret {name}"),
-			Err(e) => tracing::warn!("could not remove secret {name}: {e}"),
+			Ok(()) => {
+				tracing::info!("removed secret {name}");
+				true
+			}
+			Err(e) => {
+				tracing::warn!("could not remove secret {name}: {e}");
+				false
+			}
 		}
 	}
+}
+
+/// One summary log line that lists the podup-created secrets removed on
+/// `down`. The per-secret `info` lines stay as the per-record audit;
+/// the summary is the line the operator actually searches for when
+/// the teardown finishes. Empty lists are silent — a `down` with no
+/// declared secrets composes cleanly with the rest of the teardown
+/// without adding noise.
+fn log_removed_secrets(project: &str, removed: &[String]) {
+	if removed.is_empty() {
+		return;
+	}
+	tracing::info!(
+		"removed {} podup-created secret(s) for project {project}: {}",
+		removed.len(),
+		removed.join(", ")
+	);
 }
 
 #[cfg(test)]
