+++ internal/update/install.rs
@@ -23,7 +23,7 @@ pub fn platform_asset() -> Option<&'static str> {
 
 /// Map an OS/ARCH pair to its release asset name. Split out from
 /// [`platform_asset`] so the full matrix is testable without the host's values.
-fn asset_for(os: &str, arch: &str) -> Option<&'static str> {
+pub(crate) fn asset_for(os: &str, arch: &str) -> Option<&'static str> {
 	match (os, arch) {
 		("linux", "x86_64") => Some("podup-linux-x86_64"),
 		("linux", "aarch64") => Some("podup-linux-arm64"),
@@ -66,28 +66,91 @@ pub fn install_binary(new_bytes: &[u8], expected_version: &str) -> crate::Result
 			exe.display()
 		))
 	})?;
-	// Keep the current binary in memory so a failed self-test can roll back. The
-	// signature already proves the new bytes are authentic; the self-test guards
-	// the install mechanics (a partial write, an arch/ABI mismatch the asset name
-	// didn't catch) and pins the reported version against the resolved tag.
-	let backup = std::fs::read(&target).ok();
-	install_at(&target, new_bytes)?;
-	if let Err(e) = self_test(&target, expected_version) {
-		return match backup {
-			Some(old) => {
-				install_at(&target, &old)?;
-				Err(ComposeError::Update(format!(
-					"the updated binary failed its self-test ({e}); rolled back to the \
-					 previous version"
-				)))
-			}
-			None => Err(ComposeError::Update(format!(
-				"the updated binary failed its self-test ({e}) and no backup was \
-				 available to roll back"
-			))),
-		};
+	// Move the current binary aside BEFORE the swap so a failed self-test
+	// always has a rollback target on disk. The previous code read the
+	// bytes into a `Vec<u8>` and re-ran the install on rollback, which is
+	// a one-shot rollback: a kill between the swap and the self-test
+	// drops the in-memory copy and the user is left without a working
+	// binary. The on-disk `.old` sibling survives any kill in the window
+	// and the self-test reads from it (which is the whole point of having
+	// it on disk rather than in memory).
+	let backup = move_target_aside(&target)?;
+	if let Err(e) = install_at(&target, new_bytes) {
+		// The swap itself failed: restore the old binary before reporting.
+		let _ = restore_from_backup(&target, &backup);
+		return Err(e);
 	}
+	if let Err(e) = self_test(&target, expected_version) {
+		// Roll back to the original binary. The backup is on disk, so a
+		// kill inside `restore_from_backup` would still leave the user
+		// with a recoverable `.old` to fall back on.
+		restore_from_backup(&target, &backup)?;
+		return Err(ComposeError::Update(format!(
+			"the updated binary failed its self-test ({e}); rolled back to the \
+			 previous version"
+		)));
+	}
+	// Best-effort remove the `.old`; a kill here just leaves a recoverable
+	// sibling that the next run will tidy up.
+	let _ = std::fs::remove_file(&backup);
 	Ok(target)
+}
+
+/// Move the existing target to a sibling `.old` path so the swap and the
+/// self-test can both find a recoverable copy of the previous binary. The
+/// `.old` extension matches the Windows updater's path; the same name is
+/// used on Unix so a human inspecting the install directory sees one
+/// consistent leftover shape, and the same next-run cleanup applies.
+pub(crate) fn move_target_aside(target: &Path) -> crate::Result<PathBuf> {
+	let backup = target.with_extension("old");
+	if backup == *target {
+		return Err(ComposeError::Update(format!(
+			"refusing to clobber the target with a sibling of the same path: {}",
+			backup.display()
+		)));
+	}
+	// Drop any stale leftover from a prior interrupted install. The
+	// updater's next-run cleanup covers this too, but doing it here keeps
+	// the rename atomic from the caller's perspective.
+	if let Err(e) = std::fs::remove_file(&backup) {
+		// ENOENT is fine; anything else is worth surfacing.
+		if e.kind() != std::io::ErrorKind::NotFound {
+			return Err(ComposeError::Update(format!(
+				"cannot remove stale backup {}: {e}",
+				backup.display()
+			)));
+		}
+	}
+	if target.exists() {
+		std::fs::rename(target, &backup).map_err(|e| {
+			ComposeError::Update(format!(
+				"cannot move the current binary aside before the swap ({} -> {}): {e}",
+				target.display(),
+				backup.display()
+			))
+		})?;
+	}
+	Ok(backup)
+}
+
+/// Restore the `.old` sibling back onto `target`. Best-effort: the rollback
+/// path reports the original error to the user regardless, but a failure
+/// here is logged at debug so the next-run cleanup can pick up where this
+/// left off (the `.old` is still on disk and the self-test has already
+/// failed).
+pub(crate) fn restore_from_backup(target: &Path, backup: &Path) -> crate::Result<()> {
+	if !backup.exists() {
+		// Nothing to restore from (the install was a fresh deploy onto a
+		// path that did not exist before). The swap that already ran
+		// stands.
+		return Ok(());
+	}
+	std::fs::rename(backup, target).map_err(|e| {
+		ComposeError::Update(format!(
+			"failed to roll back to the previous binary from {}: {e}",
+			backup.display()
+		))
+	})
 }
 
 /// How long to keep retrying a spawn that reports ETXTBSY.
@@ -107,7 +170,7 @@ const TEXT_FILE_BUSY_BUDGET: std::time::Duration = std::time::Duration::from_sec
 /// that window. Treating it as a failed self-test rolls back a signed, verified,
 /// perfectly good update over a race that resolves in milliseconds, and tells
 /// the user their new version is broken.
-fn spawn_version_probe(target: &Path) -> std::io::Result<std::process::Child> {
+pub(crate) fn spawn_version_probe(target: &Path) -> std::io::Result<std::process::Child> {
 	use std::process::{Command, Stdio};
 
 	let probe = || {
@@ -143,7 +206,7 @@ fn spawn_version_probe(target: &Path) -> std::io::Result<std::process::Child> {
 /// a `String` the errno is gone, and matching on the text would break under any
 /// locale or libc that words it differently.
 #[cfg(unix)]
-fn is_text_file_busy(e: &std::io::Error) -> bool {
+pub(crate) fn is_text_file_busy(e: &std::io::Error) -> bool {
 	// `ExecutableFileBusy` is the named form; the raw errno is compared too so
 	// this holds on a toolchain where the mapping differs.
 	e.kind() == std::io::ErrorKind::ExecutableFileBusy || e.raw_os_error() == Some(libc::ETXTBSY)
@@ -153,7 +216,7 @@ fn is_text_file_busy(e: &std::io::Error) -> bool {
 /// invoking `--version`, bounded by a timeout so a hung binary can't wedge the
 /// updater. The version check closes the rollback window: a replayed older
 /// (signed) release fails here and is rolled back.
-fn self_test(target: &Path, expected_version: &str) -> crate::Result<()> {
+pub(crate) fn self_test(target: &Path, expected_version: &str) -> crate::Result<()> {
 	use std::io::Read;
 	use std::time::{Duration, Instant};
 
@@ -234,7 +297,7 @@ pub fn install_at(target: &Path, new_bytes: &[u8]) -> crate::Result<()> {
 
 /// Write the new bytes to `tmp`, copy `target`'s permission bits (default 0755
 /// on Unix when the target does not yet exist), and flush to disk.
-fn write_temp(tmp: &Path, new_bytes: &[u8], target: &Path) -> crate::Result<()> {
+pub(crate) fn write_temp(tmp: &Path, new_bytes: &[u8], target: &Path) -> crate::Result<()> {
 	// Create the temp file private (0600) on Unix so the new binary's bytes are
 	// never world-readable in a shared directory (e.g. /usr/local/bin) during the
 	// window before the target's mode is applied. `File::create` honours the
@@ -300,27 +363,46 @@ fn write_temp(tmp: &Path, new_bytes: &[u8], target: &Path) -> crate::Result<()> 
 /// Atomically move `tmp` onto `target`. Unix replaces the inode directly;
 /// Windows renames the in-use file aside first.
 #[cfg(not(windows))]
-fn swap_into_place(tmp: &Path, target: &Path) -> crate::Result<()> {
+pub(crate) fn swap_into_place(tmp: &Path, target: &Path) -> crate::Result<()> {
 	std::fs::rename(tmp, target).map_err(|e| rename_error(e, target))
 }
 
 #[cfg(windows)]
-fn swap_into_place(tmp: &Path, target: &Path) -> crate::Result<()> {
+pub(crate) fn swap_into_place(tmp: &Path, target: &Path) -> crate::Result<()> {
 	// A running .exe cannot be overwritten, but it can be renamed. Move it aside,
 	// put the new binary in place, then best-effort delete the old one (it may
 	// still be locked while running - if so, it is removed at the start of the
 	// next updater run by `cleanup_stale_backup`).
+	//
+	// The L5 swap path
+	// (`install_binary` → `move_target_aside` → `install_at`) means the target
+	// may already be absent when this function is called: the caller has moved
+	// it to `.old` so the self-test has a recoverable copy on disk. Track
+	// whether *this* call created the backup so the cleanup at the end only
+	// touches the file we created, and the rename error path only tries to
+	// restore from a backup we own.
 	let backup = target.with_extension("old");
-	let _ = std::fs::remove_file(&backup);
-	if target.exists() {
+	let target_existed = target.exists();
+	if target_existed {
+		// Drop any stale leftover before re-creating it. Only meaningful when
+		// we are about to be the one to create a fresh `.old`.
+		let _ = std::fs::remove_file(&backup);
 		std::fs::rename(target, &backup).map_err(|e| rename_error(e, target))?;
 	}
 	if let Err(e) = std::fs::rename(tmp, target) {
-		// Roll back so the user is not left without a binary.
-		let _ = std::fs::rename(&backup, target);
+		// Roll back so the user is not left without a binary — but only when
+		// we created the backup in this call. A pre-existing `.old` (the L5
+		// swap path) is the caller's responsibility to restore.
+		if target_existed {
+			let _ = std::fs::rename(&backup, target);
+		}
 		return Err(rename_error(e, target));
 	}
-	let _ = std::fs::remove_file(&backup);
+	// Only clean up the backup if we created it in this call. A pre-existing
+	// `.old` belongs to the caller, who decides when (and whether) to drop it.
+	if target_existed {
+		let _ = std::fs::remove_file(&backup);
+	}
 	Ok(())
 }
 
@@ -340,7 +422,7 @@ pub(crate) fn cleanup_stale_backup() {
 
 /// Turn a rename failure into an actionable error, calling out the common
 /// permission case (system install dirs need elevation).
-fn rename_error(e: std::io::Error, target: &Path) -> ComposeError {
+pub(crate) fn rename_error(e: std::io::Error, target: &Path) -> ComposeError {
 	if e.kind() == std::io::ErrorKind::PermissionDenied {
 		ComposeError::Update(format!(
 			"permission denied writing {}; re-run with elevated privileges \
@@ -356,314 +438,4 @@ fn rename_error(e: std::io::Error, target: &Path) -> ComposeError {
 }
 
 #[cfg(test)]
-mod tests {
-	use super::*;
-
-	#[test]
-	fn platform_asset_matches_known_targets() {
-		// Whatever host runs the tests, the asset (if any) must be one of the
-		// release matrix names.
-		if let Some(asset) = platform_asset() {
-			assert!(asset.starts_with("podup-"));
-		}
-	}
-
-	#[test]
-	fn platform_asset_covers_every_release_target() {
-		// Pins the OS/ARCH → asset mapping to the full `release.yml` build
-		// matrix so a newly added prebuilt (or a dropped arm) is caught here
-		// instead of failing self-update silently in the field.
-		let expected = [
-			(("linux", "x86_64"), "podup-linux-x86_64"),
-			(("linux", "aarch64"), "podup-linux-arm64"),
-			(("macos", "aarch64"), "podup-darwin-arm64"),
-			(("macos", "x86_64"), "podup-darwin-x86_64"),
-			(("windows", "x86_64"), "podup-windows-x86_64.exe"),
-			(("windows", "aarch64"), "podup-windows-arm64.exe"),
-		];
-		for ((os, arch), asset) in expected {
-			assert_eq!(
-				asset_for(os, arch),
-				Some(asset),
-				"self-update mapping drifted for {os}/{arch}"
-			);
-		}
-	}
-
-	#[test]
-	fn install_at_replaces_contents() {
-		let dir = tempfile::tempdir().unwrap();
-		let target = dir.path().join("podup");
-		std::fs::write(&target, b"old version").unwrap();
-
-		install_at(&target, b"new version").unwrap();
-		assert_eq!(std::fs::read(&target).unwrap(), b"new version");
-	}
-
-	/// A special bit on the target is never propagated onto the new binary.
-	///
-	/// `write_temp` copies the target's permissions so an install keeps whatever
-	/// mode the operator chose, and masks with `& 0o777` on the way. Without the
-	/// mask, a target that had been made setuid — by tampering, or by an
-	/// operator who did it on purpose once — would hand the freshly installed
-	/// podup the same bit, on a binary that has just been fetched over the
-	/// network. That is a privilege-escalation footgun, and it is the one
-	/// property of this function a test can actually observe.
-	///
-	/// The other three are window guards and cannot be reached in process: the
-	/// 0600 create mode is overwritten by this very copy before the function
-	/// returns, and `O_EXCL`/`O_NOFOLLOW` close a race between the unlink above
-	/// and the open. Their comments say so where they live.
-	#[cfg(unix)]
-	#[test]
-	fn write_temp_never_propagates_a_special_bit_from_the_target() {
-		use std::os::unix::fs::PermissionsExt;
-		let dir = tempfile::tempdir().unwrap();
-		let target = dir.path().join("podup");
-		std::fs::write(&target, b"old").unwrap();
-		std::fs::set_permissions(&target, std::fs::Permissions::from_mode(0o4755)).unwrap();
-		// Only meaningful if the filesystem kept the bit; some do not.
-		let target_mode = std::fs::metadata(&target).unwrap().permissions().mode();
-		if target_mode & 0o4000 == 0 {
-			return;
-		}
-
-		let tmp = dir.path().join("podup.tmp");
-		super::write_temp(&tmp, b"freshly downloaded", &target).unwrap();
-
-		let mode = std::fs::metadata(&tmp).unwrap().permissions().mode();
-		assert_eq!(
-			mode & 0o7000,
-			0,
-			"a special bit rode from the target onto the new binary: {mode:o}"
-		);
-		assert_eq!(
-			mode & 0o777,
-			0o755,
-			"the ordinary permission bits should still be carried over: {mode:o}"
-		);
-	}
-
-	#[test]
-	fn install_at_creates_when_absent() {
-		let dir = tempfile::tempdir().unwrap();
-		let target = dir.path().join("podup");
-		install_at(&target, b"fresh").unwrap();
-		assert_eq!(std::fs::read(&target).unwrap(), b"fresh");
-	}
-
-	#[cfg(unix)]
-	#[test]
-	fn install_at_preserves_executable_mode() {
-		use std::os::unix::fs::PermissionsExt;
-		let dir = tempfile::tempdir().unwrap();
-		let target = dir.path().join("podup");
-		std::fs::write(&target, b"old").unwrap();
-		std::fs::set_permissions(&target, std::fs::Permissions::from_mode(0o755)).unwrap();
-
-		install_at(&target, b"new").unwrap();
-		let mode = std::fs::metadata(&target).unwrap().permissions().mode();
-		assert_eq!(mode & 0o777, 0o755);
-	}
-
-	#[cfg(unix)]
-	#[test]
-	fn install_at_strips_setuid_from_target_mode() {
-		use std::os::unix::fs::PermissionsExt;
-		let dir = tempfile::tempdir().unwrap();
-		let target = dir.path().join("podup");
-		std::fs::write(&target, b"old").unwrap();
-		// A tampered/setuid target must not propagate its special bits onto the
-		// freshly installed binary.
-		std::fs::set_permissions(&target, std::fs::Permissions::from_mode(0o4755)).unwrap();
-
-		install_at(&target, b"new").unwrap();
-		let mode = std::fs::metadata(&target).unwrap().permissions().mode();
-		assert_eq!(mode & 0o7000, 0, "setuid/setgid/sticky must be stripped");
-		assert_eq!(mode & 0o777, 0o755);
-	}
-
-	#[test]
-	fn install_at_leaves_no_temp_files() {
-		let dir = tempfile::tempdir().unwrap();
-		let target = dir.path().join("podup");
-		install_at(&target, b"data").unwrap();
-		let leftovers: Vec<_> = std::fs::read_dir(dir.path())
-			.unwrap()
-			.filter_map(|e| e.ok())
-			.filter(|e| e.file_name().to_string_lossy().contains("update-"))
-			.collect();
-		assert!(leftovers.is_empty(), "temp file left behind");
-	}
-
-	#[test]
-	fn install_at_fails_when_target_dir_is_missing() {
-		// A target whose parent directory does not exist must fail (the sibling
-		// temp cannot be created) and must not leave anything behind.
-		let dir = tempfile::tempdir().unwrap();
-		let missing = dir.path().join("no-such-subdir");
-		let target = missing.join("podup");
-		assert!(install_at(&target, b"data").is_err());
-		assert!(!missing.exists(), "must not create the missing parent dir");
-	}
-
-	/// Write an executable stub script and return its path.
-	#[cfg(unix)]
-	fn write_stub(dir: &Path, name: &str, body: &str) -> PathBuf {
-		use std::os::unix::fs::PermissionsExt;
-		let p = dir.join(name);
-		std::fs::write(&p, body).unwrap();
-		std::fs::set_permissions(&p, std::fs::Permissions::from_mode(0o755)).unwrap();
-		p
-	}
-
-	#[cfg(unix)]
-	#[test]
-	fn self_test_passes_for_a_zero_exit_and_fails_otherwise() {
-		let dir = tempfile::tempdir().unwrap();
-		// A binary that exits 0 and reports the expected version passes; a
-		// non-zero exit fails.
-		let ok = write_stub(
-			dir.path(),
-			"ok",
-			"#!/bin/sh\necho \"podup 9.9.9\"\nexit 0\n",
-		);
-		let bad = write_stub(dir.path(), "bad", "#!/bin/sh\nexit 1\n");
-		assert!(self_test(&ok, "9.9.9").is_ok());
-		assert!(self_test(&bad, "9.9.9").is_err());
-		// A non-executable / missing target is a spawn error, not a panic.
-		assert!(self_test(&dir.path().join("nope"), "9.9.9").is_err());
-	}
-
-	/// The classification that silently did not work.
-	///
-	/// A real executable held open for writing cannot be run — the kernel
-	/// returns ETXTBSY — so this produces the genuine errno rather than a
-	/// hand-built error, and asserts the predicate the retry depends on. The
-	/// previous version of this check ran on an error already formatted into a
-	/// `String`, where the errno no longer exists: it returned false every time,
-	/// the retry never fired, and the flake it was written to prevent stayed.
-	///
-	/// Linux only, deliberately. Whether a given kernel refuses to exec a file
-	/// held open for writing — and for a script, whether the check lands on the
-	/// script or on its interpreter — is that kernel's business, verified on
-	/// Linux. Asserting it elsewhere tests the platform rather than the
-	/// classifier, and writing the test so it passes vacuously where ETXTBSY
-	/// never fires would repeat the mistake this whole change is about. The
-	/// retry itself stays on every Unix: it is a no-op where the errno does not
-	/// occur.
-	#[cfg(target_os = "linux")]
-	#[test]
-	fn a_binary_open_for_writing_is_classified_as_text_file_busy() {
-		let dir = tempfile::tempdir().unwrap();
-		let target = dir.path().join("held");
-		std::fs::copy("/bin/sh", &target).expect("/bin/sh is copyable");
-		{
-			use std::os::unix::fs::PermissionsExt;
-			std::fs::set_permissions(&target, std::fs::Permissions::from_mode(0o755)).unwrap();
-		}
-		// Held across the spawn on purpose; dropping it closes the window.
-		let _writer = std::fs::OpenOptions::new()
-			.write(true)
-			.open(&target)
-			.unwrap();
-
-		let err = std::process::Command::new(&target)
-			.arg("-c")
-			.arg("exit 0")
-			.spawn()
-			.expect_err("a binary open for writing must not be executable");
-		assert!(
-			is_text_file_busy(&err),
-			"ETXTBSY must be recognised from the io::Error itself, got {err:?}"
-		);
-	}
-
-	#[cfg(unix)]
-	#[test]
-	fn self_test_rejects_a_version_mismatch() {
-		let dir = tempfile::tempdir().unwrap();
-		// A genuinely-signed but *older* replayed release exits 0 yet reports the
-		// wrong version — the rollback gate must reject it.
-		let p = write_stub(
-			dir.path(),
-			"older",
-			"#!/bin/sh\necho \"podup 1.0.0\"\nexit 0\n",
-		);
-		let err = self_test(&p, "9.9.9").unwrap_err();
-		let msg = format!("{err}");
-		assert!(msg.contains("rollback"), "{msg}");
-		// A `v`-prefixed report still matches its unprefixed expectation.
-		let v = write_stub(
-			dir.path(),
-			"vprefixed",
-			"#!/bin/sh\necho \"podup v9.9.9\"\nexit 0\n",
-		);
-		assert!(self_test(&v, "9.9.9").is_ok());
-	}
-
-	#[test]
-	fn require_platform_asset_is_consistent() {
-		match (platform_asset(), require_platform_asset()) {
-			(Some(a), Ok(b)) => assert_eq!(a, b),
-			(None, Err(_)) => {}
-			_ => panic!("platform_asset and require_platform_asset disagree"),
-		}
-	}
-
-	#[test]
-	fn rename_error_calls_out_permission_and_generic_cases() {
-		let target = Path::new("/usr/local/bin/podup");
-		// A permission error nudges the user toward elevation.
-		let perm = rename_error(
-			std::io::Error::from(std::io::ErrorKind::PermissionDenied),
-			target,
-		);
-		match perm {
-			ComposeError::Update(msg) => {
-				assert!(msg.contains("permission denied"));
-				assert!(msg.contains("sudo"));
-			}
-			_ => panic!("expected an Update error"),
-		}
-		// Any other error reports the underlying failure verbatim.
-		let other = rename_error(std::io::Error::other("disk full"), target);
-		match other {
-			ComposeError::Update(msg) => {
-				assert!(msg.contains("failed to install update"));
-				assert!(msg.contains("disk full"));
-			}
-			_ => panic!("expected an Update error"),
-		}
-	}
-
-	#[cfg(windows)]
-	#[test]
-	fn cleanup_stale_backup_removes_a_leftover_old_file() {
-		// Simulates the case swap_into_place leaves behind: an `.old` sibling of
-		// the running executable that its own best-effort delete could not
-		// remove because the old process still held it open. The next updater
-		// run calls this once nothing holds the file anymore, and it must go.
-		let exe = std::env::current_exe().unwrap();
-		let backup = exe.with_extension("old");
-		std::fs::write(&backup, b"leftover backup").unwrap();
-
-		cleanup_stale_backup();
-
-		assert!(!backup.exists(), "the stale .old backup must be removed");
-	}
-
-	#[cfg(windows)]
-	#[test]
-	fn cleanup_stale_backup_is_a_no_op_without_a_leftover() {
-		// No `.old` file present is the common case (a normal run, or a
-		// platform that never took the Windows swap path) - must not error.
-		let exe = std::env::current_exe().unwrap();
-		let backup = exe.with_extension("old");
-		let _ = std::fs::remove_file(&backup);
-
-		cleanup_stale_backup();
-
-		assert!(!backup.exists());
-	}
-}
+mod tests;
